# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DreamPay API",
        "description": "High-integrity financial API for Santri Market Day ecosystem. Built with Laravel and secured with Sanctum.",
        "contact": {
            "name": "DreamPay Team",
            "email": "admin@dreampay.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.dreampay.id/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/split-bill": {
            "post": {
                "tags": [
                    "Split Bill"
                ],
                "summary": "Create a new Split Bill",
                "operationId": "03ca0bc2ec34442c6e57eedbd65b3144",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "total_amount",
                                    "description",
                                    "participants"
                                ],
                                "properties": {
                                    "total_amount": {
                                        "type": "number",
                                        "example": 60000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Dinner split"
                                    },
                                    "participants": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 30000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Split Bill created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Split Bill created successfully"
                                        },
                                        "split_bill": {
                                            "$ref": "#/components/schemas/SplitBill"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/split-bill/pay/{memberId}": {
            "post": {
                "tags": [
                    "Split Bill"
                ],
                "summary": "Pay a specific Split Bill member share",
                "operationId": "ce2ceb015f68d1308c910086c25c4cba",
                "parameters": [
                    {
                        "name": "memberId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pin"
                                ],
                                "properties": {
                                    "pin": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment successful"
                                        },
                                        "current_balance": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid PIN"
                    },
                    "404": {
                        "description": "Bill not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/split-bill/created": {
            "get": {
                "tags": [
                    "Split Bill"
                ],
                "summary": "Get bills created by me",
                "operationId": "51f3d7ed693d6c0353dbd06d0ca82da7",
                "responses": {
                    "200": {
                        "description": "List of created bills",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SplitBill"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/split-bill/incoming": {
            "get": {
                "tags": [
                    "Split Bill"
                ],
                "summary": "Get bills I need to pay",
                "operationId": "a5d604ad9699e50cce2a1d3720a49597",
                "responses": {
                    "200": {
                        "description": "List of incoming bills",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SplitBillMember"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SplitBill": {
                "title": "Split Bill",
                "description": "Split Bill model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "owner_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "total_amount": {
                        "type": "number",
                        "example": 60000
                    },
                    "description": {
                        "type": "string",
                        "example": "Dinner split"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    }
                },
                "type": "object"
            },
            "SplitBillMember": {
                "title": "Split Bill Member",
                "description": "Split Bill Member model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "split_bill_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "amount": {
                        "type": "number",
                        "example": 20000
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}